feat(M): Move supabase logic to lib folder
