+++ frontend/package.json
@@ -52,6 +52,7 @@
     "next-intl": "^4.1.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-hook-form": "^7.56.4",
     "swr": "^2.3.3",
     "tailwind-merge": "^3.3.0",
     "zod": "^3.25.42"
