+++ frontend/src/app/[locale]/(auth)/signout/route.ts
@@ -1,4 +1,4 @@
-import { createClient } from '@/utils/supabase/server';
+import { createClient } from '@/lib/supabase/server';
 import { HttpStatusCode } from 'axios';
 import { revalidatePath } from 'next/cache';
 import { NextRequest, NextResponse } from 'next/server';
